requirements.txt
# ══════════════════════════════════════════════════════════════
# Desktop Ponies v1.69 — Full dependency list
# Install all:  pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════

# ── Core GUI ──────────────────────────────────────────────────
PyQt5>=5.15.0
Pillow>=10.0.0

# ── Wake word detection ───────────────────────────────────────
pvporcupine>=3.0.0
pvrecorder>=1.2.0

# ── Speech-to-text (STT) ─────────────────────────────────────
SpeechRecognition>=3.10.0       # Google Web Speech API + mic input
pyaudio>=0.2.13

# ── Speaker verification (voice filter) ──────────────────────
resemblyzer>=0.1.3
scipy>=1.10.0
librosa>=0.10.0

# ── Text-to-speech (TTS) ─────────────────────────────────────
elevenlabs>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1

# ── LLM providers ────────────────────────────────────────────
anthropic>=0.20.0
openai>=1.14.0

# ── Vision / screen ──────────────────────────────────────────
opencv-python>=4.8.0
mss>=9.0.0

# ── CLIP watch mode (optional — scene understanding + OCR) ───
transformers>=4.30.0
torch>=2.0.0
winocr>=0.0.6                   # Windows-only OCR

# ── Desktop automation ────────────────────────────────────────
pyautogui>=0.9.54
pygetwindow>=0.0.9
pywin32>=306                    # win32gui, win32con, etc.

# ── Config / utilities ────────────────────────────────────────
PyYAML>=6.0.1
numpy>=1.24.0
